requirements.txt
Django==2.0.6
django-ckeditor==5.5.0
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
Pillow==5.1.0
pytz==2018.4
sqlparse==0.2.4
